feat: Native API Logging Improvement [DET-3316]
